Add links to source and new issue page
I don't feel great about linking to GitHub but that's because I don't
feel great about *using* GitHub for source hosting and bug tracking. On
the other hand, I think it's useful to link to both since the people who
will be trying this are likely going to be developers right now. In the
future I'll probably revert this change.