Change link redirect to center messsage footer
Previously when you liked a post it would redirect you back to the
previous page with the top of the post at the top of the screen. This
was jarring and weird and I didn't like it at all.

This change makes it so that the footer of the message you liked
appears in the center of the screen, which is (in my experience) much
closer to where it usually is when you click the like button.

Before: when you click the like button the scroll resets so the top of the
message is at the top of the screen.

After: when you click the like button the scroll resets so that the footer of
the message is in the center of the screen.

Wanted: a way to redirect back to the original page *without* resetting the
scroll position.