Fix comments and rename 'reply all' to 'comment'
This changes some phrasing for clarity and adds some helpful
explanations when you're publishing a comment or a reply.

This also fixes comments on replies, which were previously just being
added as a sibling reply. This doesn't really matter because it has the
same layout in the UI and it's also very rare, but it allows us to have
separate threads for each reply.