Solve inconsistent widths between pages
Problem: The `<section>` elements on some pages have different sizes,
which means that the sidebar is doing wonky stuff.

Solution: Make the sections all have a consistent width. Note that
there's still a few pixels of jank on the 'search' page, I'm not sure
why that's happening. The element inspector is reporting that everything
is the same width but that isn't true.