Remove inline CSS to tighten CSP
Problem: We were using inline CSS to show off the theme on the 'meta'
page, which meant that we couldn't tighten the CSP. This meant that
if you could figure out a CSS injection you'd be able to change how
Oasis looks, which could probably be malicious.

Solution: Add the CSS to `style.css` and remove that attack surface
area. Removing this inline CSS is a best practice, and while I don't
think a CSS injection is a super realistic threat there's no reason to
leave open holes in our CSP.