Add basic following page
Problem: We don't always want to see all of the messages on our
computer, sometimes we just want to see messages from the people we're
explicitly following. The 'Popular' and 'Latest' pages don't help with
that.

Solution: Add a super basic page that just shows the latest messages
from the people you're explicitly following.