Improve readme with more details about Oasis
Problem: The previous readme didn't really have very much information
about what Oasis is or what we want it to be. It wasn't very enticing.

Solution: Zoom out a bit and give a top-3 list of features that
distinguishes Oasis from other social networking applications and SSB
applications in general.