Add AUTHORS and MAINTAINERS files
Problem: It's hard to know who has worked on the project and who is
maintaining the project, for security and code of conduct issues.

Solution: Add AUTHORS and MAINTAINERS files to track who has contributed
code and mixed in their copyright (woo!) and keep track of who is
committed to maintaining the project and merging patches.