Avoid making installer build from source
Problem: The previous approach of having the installer run `npm install`
was hazardous at best and didn't work well offline.

Solution: Package all of the source code with the application and use
the installer to ensure that it works on the target device. The
installer should ensure that we can `require()` the module, and if that
doesn't work then it should choose between two options:

- If the module is optional, `rm -rf` it and don't worry about it.
- If the module is mandatory, try running `npm install` with some flags
  to ensure that we stay offline and print the absolute minimal output.
  This should never happen, because we should never be packaging Oasis
  for architectures and platforms that we don't have prebuilds for, but
  it's a fine fallback behavior for experimental hackery.