Remove debug output and add comments
Problem: The previous code was printing debug output so that I could
ensure that my argument checking worked correctly, but we don't want to
distribute a binary that echoes back your arguments to you.

Solution: Remove the debug output and add lots of comments to document
all of my misunderstandings of how the current code works. :~)