Replace "metric CSS" with --size-N
Problem: The "metric CSS" was kind of clever but mostly just confusing.

Solution: Use `--size-N` for the size, where the size is `2^n`. Double
size is `--size-1`, half size is `--size--1` (negative one). The
negatives kind of suck but I couldn't think of a better solution that
didn't give up the flexibility, brevity, or the ability to sort
alphabetically.

Fixes: https://github.com/fraction/oasis/issues/100