Add super simple same-as implementation
Problem: Most of us have lots of feeds and there's no way to indicate
which feeds we want to see notifications for.

Solution: Start reading the `sameAs` profile property, and if it's an
array with feed IDs then also look up mentions for those feed IDs. Note
that this commit doesn't introduce a way to add same-as in the UI, so
you ~~have~~ get to use the custom message publish feature to make a
message like:

```json
{
  "type": "about",
  "about": "<your current feed id>",
  "sameAs": [
    "<another feed id>"
  ]
}
```

The `sameAs` array can have as many items as you'd like, although I've
only actually tested with two.