Use stricter regex for @mention match
Problem: You can't publish email addresses or user@host combos or
anything like that because the regex is super liberal in what it
matches.

Solution: Ensure that the first word in a message is a mention *or*
ensure that there's whitespace before a mention. This also adds a
regex101 link (in lieu of proper testing, for now) which makes this
easier to test and track over time.

See-also: https://github.com/fraction/oasis/issues/466